refactor(simulator): extract OOP/class runtime into simulator/oop.rs (Step 19)
Move the ClassInstance struct and the class-runtime methods (class_prop_width,
class_static_get, class_static_set, exec_method_call, class_prop_width_of,
exec_method_in_class_hierarchy, class_prop_type) into a dedicated
simulator/oop.rs submodule. mod.rs shrinks by ~1080 lines. Struct/fields/methods
made pub(crate). 1935 tests pass.
